Selection tabs tweaks.
Fix PR comments including some UI tweaks and bug fixes.

Hide no annotations/notes message when annotations are still loading.
When the annotations are still loading, a 0 annotations message shows for a flash of a
second before the list loads. Avoid this by only displaying the 0 state after the
annotations have finished loading.

If selected annotation/note is deleted, default selected tab to annotation.
If a user lands on a direct link for a deleted annotation, the button to take the user
back to the list of annotations/notes should by default be for annotations.

Fix broken standalone page for notes.
selectedTab was set to annotations by default, even for standalone pages.
selectedTab should be undefined for a standalone page because there are no tabs to select from.
select the annotations tab by default for the sidebar alone.