Add fuzz tests for anchoring
Add a fuzz test that generates a randomized DOM structure and randomized
Range within that DOM structure.

It then attempts to describe the range using the HTML anchoring module
and re-anchor each of the generated selectors using the HTML anchoring
module.

The test currently triggers numerous exceptions during describing and
anchoring. Each of the known failure cases is whitelisted by a regex on
the exception message and stacktrace. As issues are fixed, the
corresponding entries can be removed from the whitelist.