Remove requirement to pass search scope property down from app to child controllers
In preparation for converting the top level app controller and content
areas of the app into components which will not inherit scope
properties, remove the need for the `search` property to be passed down
on the scope.

Just centralize the logic for retrieving and handling changes to the
search query in the top-level app controller.