Update to Angular 1.6.9
The main change that affects us is that component inputs/bindings are no
longer available as controller properties in the constructor by default.
Instead this logic should go in a controller's `$onInit` method.

`$compileProvider.preAssignBindingsEnabled` is a compatibility aid
which preserves the Angular <= 1.5 behavior. It is due to be removed
in Angular 1.7.