Support finding app assets relative to boot script URL
Enable the boot script to determine the URL of app assets to load based
on the boot script's own URL. This avoids needing to hardcode these URLs
at the time the client is built and makes it "relocatable".

This requires browser support for `document.currentScript`. In practice
this only rules out IE 11 and very old versions of Chrome (< 29) and
Safari (< 8).