Convert new design part of `<group-list>` to React
Convert the part of the `<group-list>` component which implements the
new design to React. In order to support switching the menu design depending on
the `community_groups` feature flag, the converted component is rendered
inside the existing component as `<group-list-v2>`. When the old design
and feature flag are removed, the `<group-list>` and `<group-list-v2>`
components can be combined back into one.

The converted component is the same as before, except that the
"New private group" link has been converted into a link for
accessibility.