Add mixin to enable focus ring only when an element has keyboard focus
In certain contexts it is desirable to disable the browser's default
outline around focused elements, but not if the focus ring is necessary
for a11y, such as when using keyboard navigation.

Provide a mixin to do this, using the :focus-visible pseudo-class, along
with a polyfill since :focus-visible is not yet widely supported (only
Chrome, behind a flag).