Convert `Excerpt` component to Preact
Convert the `Excerpt` component to Preact. Rather than convert the
existing implementation verbatim, this is a completely new implementation which
should be easier to understand and use. Callers no longer need to
provide a `contentData` input as the excerpt directly monitors the DOM
for changes using `ResizeObserver` if available or `MutationObserver` +
DOM events otherwise.

This component renders caller-provided content (ie. it accepts a `children`
prop), which is not supported by the Preact <-> Angular bridge. Therefore it was
also necessary to create components (`AnnotationBody`, `AnnotationQuote`) that
encapsulate uses of `Excerpt` and their children.