Add function to convert DOM ranges to text positions
Add a `fromRange` function which will serve as a replacement for the
`fromRange` function in `dom-anchor-text-position`.

This implementation provides better / more explicit handling of ranges that
are entirely or partly outside of the `root` element. It acts as if the
range's start and end positions were clamped to the boundaries of the
root element.