Fix `describe` failing if range starts/ends outside text layer
Rewrite the `describe` function so that it works if the passed range starts or
ends outside of the text layer. This can happen when the makes a selection that
includes the PDF.js UI at the top of the page, or if they press Ctrl/Cmd+A to
select all the visible text in the document.

 - Change the logic for finding the `.textLayer` div containing the
   selection so that it works if the range starts or ends outside of a
   text layer. A trivial way to make this happen is to press Ctrl/Cmd+A
   to create a selection

 - Add a dummy footer and header to fake PDF viewer UI to simulate the
   UI elements that PDF.js has, and use this to create a test case for
   ranges that extend beyond the PDF viewer

 - Fix an issue where the text quote selector used an ancestor of the
   text layer element as its root, allowing it to generate selectors
   that referred to text outside of the text layer