Fix an issue where annotations can sometimes anchor but fail to be highlighted
When a text position occurs at the boundary between two text nodes,
there are two ways that a DOM Range can be created which starts at
this boundary: It can start at the end of the first node or at the
beginning of the next node.

Downstream code ought to be handle either case, however this is not the
case for the `Guest#anchor` method. The problem can be reproduced using
the anchoring integration tests.

To resolve this, make sure the returned range always starts at the start of
a node and ends at the end of a node, when there is a choice.

This choice also makes sense in its own right as it makes the returned
range as "small" as possible.