Add `watch` method to store to run a callback when certain state changes
A common need in the sidebar app is to trigger work in response to
certain state in the store changing. We currently do this either via
`store.subscribe` calls where the listener manually remember the previous state
or using Angular `$scope.$watch` calls.

Add a `watch` method to the store which assists with this.