WIP - Move logic for initiating annotation fetch to annotations service
In addition to paving the way for converting the `<sidebar-content>`
component to Preact, this also speeds up the initial annotation search
by initiating it as soon as the store has been populated with the
necessary information.

There is currently a delay between groups being loaded and the
annotation fetch using `/api/search` being initiated. (TODO - Explain
where this delay comes from)