WIP - Change Guest/Sidebar relationship
Change the relationship between the `Guest` and `Sidebar` classes from
one of inheritance to one where `Sidebar` receives a reference to the
`Guest` instance.

In the process it was necessary to fix a design issue where the `Guest`
class declared a `toolbar` property which was initialized by the
`Sidebar` subclass. What now happens instead is that the `Guest` class
emits events and the `Sidebar` class handles updating the toolbar as
necessary.