Replace `isNodeInRange` with `range.intersects(node)`
Replace a custom function for testing whether a node lies within a range
with a standard DOM API which is supported by all of our target
browsers.

This change resolves an issue where `isNodeInRange` would fail if `node`
had no parent, an issue that was noticed during changes to the
highlighter implementation.