Implement `useStoreProxy` hook
Implement a new custom hook that provides access to read and update the
store in UI components. This will replace the existing `useStore` hook.

The new hook has a better API which is more ergonomic to use and should
prevent common mistakes seen when using the previous API. It returns a
proxy for the store. The UI component can use the proxy just like it would use
the store directly. The proxy however tracks and caches the results of
store selector calls. When the store updates it checks whether any of
the cached calls would return different results and if so invalidates
the cache and re-renders the component.