Use a separate CSS class for annotator CSS reset
The `annotator-frame` class applied to the sidebar's CSS container has
two functions:

 1. Style the outermost container element of the Hypothesis sidebar
 2. Apply a CSS reset to override browser defaults and, in browsers
    which don't support shadow DOM, host page styles

Once we retire support for browsers that don't support Shadow DOM, we
can change the CSS reset to be loaded "directly" within the shadow roots
created by annotator code, rather than requiring a wrapper class.

We're not ready to do that just yet, but this commit paves the way by
splitting out the CSS reset function of the `annotator-frame` class into
an `annotator-reset` class.