Avoid loading annotator UI component CSS in host page
In browsers that support shadow DOM, the annotator.css bundle does not
need to be loaded in the host page because it is loaded inside the
shadow roots of Hypothesis UI elements. Only the CSS for highlights and
PDF.js overrides is needed in the host page.

Avoiding loading the annotator.css bundle in the host page will allow us
to apply global CSS (eg. our standard CSS reset) in that bundle without
worrying about it affecting other parts of the host page.

 - Split CSS for highlight elements into separate bundle and load it in
   the host page
 - In browsers that don't support shadow DOM, continue to load the
   annotator.css bundle in the host page
 - In browsers that do support shadow DOM, don't load the annotator.css
   bundle in the host page but do preload it, since we know it will be
   needed