WIP - Implement side-by-size mode for HTML documents
We recently improved the experience of annotating PDF documents by
showing the sidebar alongside the content when there is enough
horizontal space available. This commit extends that approach to web pages,
with modifications to allow for differences between web pages and PDFs:

 - Web pages do not have a known layout and response to being resized
   that we can rely on

 - Web pages will often have peripheral content on the side which we
   should not try to squeeze into the available space when the sidebar
   is opened. Instead we should continue to allow the sidebar to cover
   it up to make more room for the main content

This is implemented by first shrinking the `<body>` by the full width of
the sidebar when it opens and then increasing the `<body>`'s width if
there is space between the "main content" of the page and the edge of
the sidebar. Whatever is in this space (eg. related stories, ads) will
then be covered up by the sidebar.

Determining what consitutes the "main content" of the page
is done by heuristics in `guessMainContentArea`. These heuristics assume
that the main content can be inferred by considering the longest
paragraphs of text on the page and looking for common vertical margins
amongst them.

Resizing the document will generally cause the content to shift. To try
and ensure the same content remains visible before and after opening the
sidebar, a `preserveScrollPosition` helper attempts to determine how
much the content shifted after the relayout and adjust the scroll
position to keep, near-enough, the same content visible. To do this, it
chooses an element near the top of the screen as an anchor and checks
how much it shifts following the resize. It then adjusts the document's
scroll position to compensate.