Add `createStoreModule` utility
Add a `createStoreModule` function for creating store modules with less
boilerplate and better type safety. The resulting modules have the same
structure and external behavior as existing modules and so individual modules
can be converted to use this utility without external changes.

`createStoreModule` reduces boilerplate primarily by combining the reducers and
action creators into a single `actions` object. Each value in this
object is a function that takes the read-only module state as an
argument, plus the arguments for a specific action, and returns the
state changes resulting from the action. Boilerplate is also reduced by
automatically generating the types of actions using a combination of the module's
namespace and the action name, eg. `toastMessages/addMessage`.

`createStoreModule` improves type safety by enabling the type of the
`state` parameter to action and selector methods to be automatically
inferred based on the type of the initial module state.