Rename `storeModule` => `createStoreModule`
This function was originally written as a helper to enable us to enable
TypeScript to generate more useful errors if a store module had the
wrong structure. However it also provides a place that we can do runtime
checks / normalization of store modules etc.

Rename the function and update the documentation to make it clear that
this is the correct way to define a store module.