Use separate channels for guest / host messages from sidebar
Use separate channels for sending messages to guests vs the host in the
sidebar, even for the common case when there is only one guest and it is
the same frame as the host.

This change makes it clear for readers which part of the annotator code
is intended to handle a particular message from the sidebar. It also
paves the way towards supporting scenarios where the host frame is not a
guest frame, such as in an ebook reader where the host frame is the book
viewer but the guest is an iframe within the reader displaying the
content of a particular chapter. In this scenario only the chapter
content should be annotatable, not the book UI.