Rewrite `getPageTextContent` to clarify sync vs async behavior
An important aspect of this function is that although it returns an
async result, it populates the cache synchronously, so that two racing
calls to `getPageTextContent` for the same page don't ask PDF.js to
extract the same text twice. Add comments to make that clear and remove
the `async` from the outer function.

 - Combine the two internal async helpers into one with no arguments.
   This eliminates some untyped parameters.

 - Move the documentation about returned text matching text layer from
   an internal comment to the function's JSDoc, since it is something
   callers need to know even if they don't care about the
   implementation.