Add support for multiple guests in bucket bar
Support displaying anchors from guests in same-origin iframes in the
bucket bar. Replace the single guest argument to the constructor with an
`addGuest` method which is called for each guest that is created. A
change was also required to the bucket-building logic to correctly
determine the location of anchors in iframes relative to the viewport
that contains the bucket bar.