Anchor quotes to best match in whole PDF
Anchor quote selectors to the best match in the whole PDF, rather than
the best match in the first-searched page which contains at least one match above
the minimum quality threshold.

This fixes an issue described where sub-optimal matches could be returned
if the position hint is no longer valid due to changes in PDF.js text rendering
between the PDF.js version in which the annotation was created and the current
PDF.js version.

A downside of this change is that anchoring a quote selector now requires
extracting the text of the entire PDF, whereas before quote anchoring typically
only needed to extract text for pages up to and including the one that contained
the quote. This can make anchoring feel slower in very long documents.

Fixes https://github.com/hypothesis/client/issues/3705