Find best match for quote selectors across all pages ignoring whitespace
As described in #3705, stopping a quote search as soon as we find an
approximate match could result in quotes anchoring to suboptimal matches
if the position selector is no longer accurate.

We want to balance finding the best match against making the search
efficient in long documents. The approach taken in this commit is to
continue to search pages in priority order, but instead of returning the
first "good enough" match (below `matchQuote`'s error threshold), return
the first exact match ignoring whitespace differences, or the nearest
match in the whole document otherwise.

We ignore whitespace differences because this is the main change between
the current version of PDF.js that we include in Hypothesis products, and
newer versions that we are trying to migrate to (see #3687). It is also
likely to be the main source of text differences between PDF.js and
other viewers.

 - Change `anchorQuote` to ignore whitespace differences when searching
   for approximate matches
 - Change `anchorQuote` to keep searching for the best match if it does
   not find a whitespace-insensitive exact match in the current page
 - Add tests for whitespace changes

Fixes https://github.com/hypothesis/client/issues/3705