Support scrolling highlights in child frames into view
When clicking on a card in the sidebar to scroll a highlight into view,
we may need to scroll not only the frame containing the highlight but
also the parent frame. This is the case in the "tall frame" example in the dev
server (http://localhost:3000/document/tall-iframe) and the new
VitalSource book reader.