Generate PDF quote selectors from extracted PDF text
Generate quote selectors from the extracted PDF text instead of text layer DOM
element. This ensures that the quote selector does not include any content
outside of the text layer (eg. whitespace in the HTML at the start or end of the
document) or content from the text layer of un-rendered pages, such as the
"Loading annotations..." placeholder that we create.

Fixes #3737