WIP - Coalesce highlights that cover the same text
When exactly the same text is highlighted multiple times, only create
one highlight element. This has several benefits:

 - It reduces the total number of DOM nodes, especially in heavily
   annotated pages, which improves performance

 - It reduces noise for screen reader users, as the text added around
   annotations in `.hypothesis-highlight::{before, after}` is read less
   often

To support this a reference count has been added to highlights via a
`refCount` property. This is incremented each time a sequence of text
nodes is highlighted and decremented when `removeHighlights` is called.
The highlight element is not removed until the reference count reaches
zero.