Execute quote searches in a web worker
Quote searches can block the UI when many annotations are anchored in very long
documents. Attempt to execute the quote search in a web worker to avoid blocking
the main thread in that scenario.

In future we might want to move additional pieces of the anchoring logic to a
web worker. The quote searching logic is a good starting point because it can
be expensive and doesn't touch the DOM.

 - Add a small new bundle that uses `match-quote.js` as an entry point.
   The `preserveEntrySignatures` setting was removed because it is no
   longer needed and it prevents creation of bundles with exports.

 - Add `runInWorker` utility that calls a function from a script in a
   web worker. The worker currently has to use a non-module script
   because Firefox doesn't support dynamically loading ES modules in workers.

 - Add `matchQuoteAsync` function to `match-quote.js` that uses
   `runInWorker` to execute a `matchQuote` call in a worker.

 - Modify `TextQuoteAnchor` class to use `matchQuoteAsync` to execute
   quote searches asynchronously in a way that does not block the UI.

 - Change tsconfig.json module configuration to enable use of
   `import.meta.url`, used by `runInWorker` to resolve the path of the
   ES module script.