WIP - Optimize anchoring of short, generic quotes
Re-work the fuzzy quote matching algorithm to handle short, generic quotes
better.

If a quote was short and very generic, this could result in a very large
number of matches which were then individually scored before being
sorted. In the associated bug [1], a 200K char document had a quote
consisting of a single space, generating 40K matches. For each of these
40K matches a score based on the prefix, suffix and position was
generated. The prefix and suffix scoring required a lot of approximate
string matching.

Re-work the algorithm so that the context, quote and suffix are always
matched against the document in a single pass.