WIP - Replace Bridge class with PortRPC
Modify PortRPC's API for registering event handlers and connecting to be
more like Bridge and MessagePort, where constructing the class,
registering method/event handlers and starting the connection are
separate steps. This separation is convenient when the MessagePort is
not available immediately but only after a discovery process.

Also parametrize PortRPC by the allowed `on` and `call` methods, similar
to Bridge.

Existing uses of Bridge have been replaced either with a single PortRPC
instance, for 1:1 connections between frames, or an array of PortRPC
instance for 1:N connections between the sidebar/host and guest frames.

TODO:

 - Decide what to do about typing method call names from Bridge
 - Remove guest connections in FrameSyncService when host frame notifies
   us that guest has gone away
 - Decide how to deal with invoking `PortRPC.call` before a connection
   is established