Re-implement `getScrollAnchor` to ignore text nodes in hidden elements
The previous implementation of `getScrollAnchor` could choose words from
a non-displayed text node (eg. inside a `<title>` tag) as the anchor.
This is because `Range.getBoundingClientRect()` for such a word would
return a DOMRect with coordinates (0, 0, 0, 0) which passed the test for
being contained within the viewport (0, 0, window.innerWidth,
window.innerHeight).

Fix this with a rewrite of `getScrollAnchor` which also improves
efficiency by skipping over text nodes in subtrees of the root element
which do not intersect the viewport.