Simplify and improve type safety of `watch` utility
Simplify the `watch` utility by making it only support a single callback to get
the watched value, along with a comparison function to use to compare the
results. To watch multiple values the callback can return an array and use a
shallow-equality comparison. If no comparison function is provided, a
strict-equality check is used.

Even though the new `compare` argument is optional, it precedes the `callback`
argument because that function is often lengthy, and calls to `watch` are more
readable if the `compare` argument comes before it.